+++ com.microsoft.mrtk.core/Editor/Inspectors/Utilities/MRTKEditorStyles.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f7e054a92532350499a6d5f06592da26
+guid: 1650069ded63dbc4fa5f154ae96ba565
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
