--- com.microsoft.mrtk.core/Editor/Inspectors/Utilities/MixedRealityInspectorUtility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b84dcb0ba4bcc1c49ae2f5bbcc31746f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
